# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItemSize {
  label: string;
  price: number;
}

export interface MenuItem {
  id: string;
  name: string;
  /** @nullable */
  nameAr?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  sizes?: MenuItemSize[] | null;
  popular?: boolean;
  /** @nullable */
  note?: string | null;
}

export interface MenuCategory {
  id: string;
  name: string;
  nameAr: string;
  icon: string;
  accent: string;
  glow: string;
  items: MenuItem[];
}

export interface MenuDocumentInput {
  categories: MenuCategory[];
}

export type MenuDocument = MenuDocumentInput & {
  binId: string;
  updatedAt?: string;
};