# lib/api-zod/src/generated/types/menuCategory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItem } from './menuItem';

export interface MenuCategory {
  id: string;
  name: string;
  nameAr: string;
  icon: string;
  accent: string;
  glow: string;
  items: MenuItem[];
}