# lib/api-zod/src/generated/types/menuDocument.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MenuDocumentInput } from './menuDocumentInput';

export type MenuDocument = MenuDocumentInput & {
  binId: string;
  updatedAt?: string;
};