# lib/api-zod/src/generated/types/menuDocumentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MenuCategory } from './menuCategory';

export interface MenuDocumentInput {
  categories: MenuCategory[];
}