# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemSize } from './menuItemSize';

export interface MenuItem {
  id: string;
  name: string;
  /** @nullable */
  nameAr?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  sizes?: MenuItemSize[] | null;
  popular?: boolean;
  /** @nullable */
  note?: string | null;
}